refactor: BUTTON_LABEL 을 const/recruitment 로 이동
