+++ frontend/src/components/MyApplicationItem/MyMissionItem/MyMissionItem.tsx
@@ -4,8 +4,7 @@ import { generatePath } from "react-router";
 import { useNavigate } from "react-router-dom";
 import { Mission } from "../../../../types/domains/recruitments";
 import { PARAM } from "../../../constants/path";
-import { MISSION_STATUS } from "../../../constants/recruitment";
-import { BUTTON_LABEL } from "../../../pages/MyApplication/MyApplication";
+import { BUTTON_LABEL, MISSION_STATUS } from "../../../constants/recruitment";
 import { formatDateTime } from "../../../utils/format/date";
 import MissionDetail from "../MissionDetail/MissionDetail";
 import ApplyButton from "../MyApplicationButtons/ApplyButton";
