+++ frontend/src/constants/recruitment.ts
@@ -11,3 +11,12 @@ export const MISSION_STATUS = {
   UNSUBMITTABLE: "UNSUBMITTABLE",
   ENDED: "ENDED",
 } as const;
+
+export const BUTTON_LABEL = {
+  BEFORE_SUBMISSION: "준비 중",
+  EDIT: "수정하기",
+  SUBMIT: "제출하기",
+  UNSUBMITTABLE: "제출불가",
+  COMPLETE: "제출완료",
+  UNSUBMITTED: "미제출",
+} as const;
