+++ frontend/src/pages/MyApplication/MyApplication.tsx
@@ -11,15 +11,6 @@ import useTokenContext from "../../hooks/useTokenContext";
 import useMissions from "./../../hooks/useMissions";
 import styles from "./MyApplication.module.css";
 
-export const BUTTON_LABEL = {
-  BEFORE_SUBMISSION: "준비 중",
-  EDIT: "수정하기",
-  SUBMIT: "제출하기",
-  UNSUBMITTABLE: "제출불가",
-  COMPLETE: "제출완료",
-  UNSUBMITTED: "미제출",
-} as const;
-
 const MyApplication = () => {
   const { token } = useTokenContext();
   const { recruitment } = useRecruitmentContext();
